errores ok, modularizado home
